requirements.txt
# Core dependencies
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.1.0  # YOLO26 support

# Computer Vision
opencv-python>=4.9.0
opencv-python-headless>=4.9.0
supervision>=0.19.0
paddlepaddle>=2.6.0
paddleocr>=2.7.0

# Tracking & Re-ID
fastreid>=1.4.0
lap>=0.4.0
filterpy>=1.4.5

# Data Processing
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
scikit-learn>=1.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Video Processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3
imageio>=2.34.0
imageio-ffmpeg>=0.4.9

# Report Generation
jinja2>=3.1.0
weasyprint>=61.0
reportlab>=4.1.0

# API & Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.2

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
